Further enum tweaks, two macros now, WRAPPED_ENUM & STRINGY_ENUM, and REGULAR_ENUM which evaluates to WRAPPED_ENUM with NDEBUG defined, else STRINGY_ENUM.  EnumNames now templated on enum type. + enum_cast<>() + continued massive re-enumeration + NodePool split into NodeStore and NodePool, pools are attached to stores before searching. + additional documentation

+++ source/glest_game/path_finder/influence_map.h
@@ -29,10 +29,10 @@ class InfluenceMap {
 
 public:
 	InfluenceMap();
-	InfluenceMap( int x, int y, int w, int h );
+	InfluenceMap(int x, int y, int w, int h);
 
-	void  clear()			{ memset( iMap, (int)0.f, sizeof(float) * width * height ); }
-	void  clear(float val)	{ std::fill_n( iMap, width*height, val ); }
+	void  clear()			{ memset(iMap, 0, sizeof(float) * width * height); }
+	void  clear(float val)	{ std::fill_n(iMap, width*height, val); }
 
 	bool  isInside(const Vec2i &pos)		const;
 	float getInfluence(const Vec2i &pos)	const;
